Refactor string grouping: replace formatter with textutil for consistency and remove unused formatter module
